+++ crates/block-producer/src/block_builder/mod.rs
@@ -95,7 +95,7 @@ impl BlockBuilder {
     }
 
     #[instrument(target = COMPONENT, skip_all, err)]
-    async fn build_block(&self, batches: &[ProvenBatch]) -> Result<(), BuildBlockError> {
+    pub async fn build_block(&self, batches: &[ProvenBatch]) -> Result<(), BuildBlockError> {
         info!(
             target: COMPONENT,
             num_batches = batches.len(),