+++ crates/block-producer/src/domain/transaction.rs
@@ -129,7 +129,7 @@ impl AuthenticatedTransaction {
     }
 }
 
-#[cfg(test)]
+#[cfg(any(test, feature = "testing"))]
 impl AuthenticatedTransaction {
     //! Builder methods intended for easier test setup.