+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 ### Changes
 
 - [BREAKING] Updated minimum Rust version to 1.84.
+- Added stress test binary with seed-store command (#621).
 - [BREAKING] `Endpoint` configuration simplified to a single string (#654).
 
 ### Enhancements